🌱 Move extracted sub-components to card-wrapper/ subdirectory
Avoids ts-null-safety CI regression: the AI quality gate scans
web/src/components/cards/*.tsx for useCardLoadingState. These helper
files are internal to CardWrapper, not standalone cards, so they
belong in a subdirectory excluded from the glob.
